Use visibility API instead of beforeunload
Using the `beforeunload` event prevents browsers from caching the page
in the page navigation cache:

https://developers.google.com/web/updates/2018/07/page-lifecycle-api#the-beforeunload-event

Page Visibility is a better alternative. This API is widely supported
except for old Android, where it's prefixed on 4.4 and unavailable on
lower versions. I'm happy to add fallbacks if any reviewers feel they're
necessary.